BACK-END FINALIZADO

+++ controllers/aut.controllers.js
@@ -9,19 +9,20 @@ const iniciarSesion = async (req, res, next) => {
   try {
     const userEncontrado = await AutModel.findOne({ email });
 
-    //* Check if user already exists
+    //* VERIFICAR SI EL USUARIO EXISTE
     if (!userEncontrado) {
       return res.status(404).json({ msg: "Usuario no encontrado" });
     }
 
-    //* Check if password is correct
+    //* REVISAR SI EL PASSWORD ES CORRECTO
     if (!(await bcrypt.compare(password, userEncontrado.password))) {
       return res.status(400).json({ msg: "Password Incorrecto" });
     }
 
-    //* ACTUALIZAR EL TOKEN
+    //* GENERAR ID PARA EL TOKEN
     const idToken = shortId.generate();
 
+    //* GENERAR TOKEN CON TODOS LOS DATOS
     const token = jwt.sign(
       {
         id: userEncontrado._id.toString(),
@@ -32,6 +33,7 @@ const iniciarSesion = async (req, res, next) => {
       { expiresIn: "90d" }
     );
 
+    //* ASIGNAR EL TOKEN AL USUARIO EN LA BD
     userEncontrado.token = token;
     await userEncontrado.save();
 
@@ -47,6 +49,7 @@ const registrarse = async (req, res, next) => {
   const errores = [];
 
   try {
+    //* VALIDACIÓNES CON EXPRESIONES REGULARES
     const ValidacionEmail = new RegExp(
       /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/,
       "g"
@@ -55,6 +58,7 @@ const registrarse = async (req, res, next) => {
     const validacionApellido = new RegExp(/^[a-zA-Z]{3,25}$/);
     const validacionPassword = new RegExp(/^([a-zA-Z0-9]){7,15}$/);
 
+    //* COMPROBAR LAS VALIDACIONES
     if (!validacionNombre.test(req.body.nombre)) {
       errores.push("El NOMBRE es inválido");
     }
@@ -72,14 +76,14 @@ const registrarse = async (req, res, next) => {
       return res.status(200).json({ msg: errores });
     }
 
-    //* Guardar usuario
+    //* GUARDAR USUARIO
     await AutModel.create(req.body);
 
     return res.status(200).json({ msg: "Registro Exitoso" });
   } catch (error) {
     const arrayErrors = [];
 
-    //! SHOW ERRORES.
+    //! PONER ERRORES EN UN ARREGLO DE ERRORES
     if (error.errors) {
       error.errors.nombre?.properties.type === "required" &&
         arrayErrors.push("El NOMBRE es obligatorio");
@@ -93,35 +97,37 @@ const registrarse = async (req, res, next) => {
 
     error.code === 11000 && arrayErrors.push("El EMAIL ya está registrado");
 
+    //! MOSTRAR LOS ERRORES ACOMULADOS
     console.log(error);
     return res.status(400).json({ msg: arrayErrors });
   }
 };
 
 const editarPerfil = async (req, res, next) => {
-  //TODO: FALTA TERMINAR COMPONENTE
-  //! AGREGAR LA VALIDACIÓN CON EL JWT => req.usuario;
   try {
-    const { id, nombre, apellido, password } = req.body;
+    const { idUsuario } = req.params;
+    const usuario = req.usuario;
+    const { nombre, apellido, password } = req.body;
 
-    const usuarioEncontrado = await AutModel.findById(id);
-
+    const usuarioEncontrado = await AutModel.findById(idUsuario);
+    //* COMPROBAR QUE EL USUARIO EXISTE
     if (!usuarioEncontrado) {
       return res.status(404).json({ msg: "Usuario no encontrado" });
     }
 
-    if (nombre) {
-      usuarioEncontrado.nombre = nombre;
+    //* COMPROBAR CON EL TOKEN QUE EL USUARIO TENGA PERMISO DE EDITAR
+    if (usuario.id !== usuarioEncontrado._id.toString()) {
+      return res
+        .status(403)
+        .json({ msg: "No tienes permiso para editar este perfil" });
     }
 
-    if (apellido) {
-      usuarioEncontrado.apellido = apellido;
-    }
+    //* ACTUALIZAR DATOS QUE EL USUARIO HAYA DADO
+    if (nombre) usuarioEncontrado.nombre = nombre;
+    if (apellido) usuarioEncontrado.apellido = apellido;
+    if (password) usuarioEncontrado.password = password;
 
-    if (password) {
-      usuarioEncontrado.password = password;
-    }
-
+    //* GUARDAR EN LA BASE DE DATOS
     await usuarioEncontrado.save();
 
     return res
@@ -137,17 +143,25 @@ const editarPerfil = async (req, res, next) => {
 };
 
 const eliminarPerfil = async (req, res, next) => {
-  //TODO: FALTA TERMINAR COMPONENTE
-  //! AGREGAR LA VALIDACIÓN CON EL JWT => req.usuario;
   try {
     const { email } = req.body;
+    const usuario = req.usuario;
 
     const usuarioEncontrado = await AutModel.findOne({ email });
 
+    //* COMPROBAR QUE EL USUARIO EXISTA
     if (!usuarioEncontrado || usuarioEncontrado.length === 0) {
       return res.status(404).json({ msg: "El usuario no existe" });
     }
 
+    //* COMPROBAR QUE EL USUARIO TENGA PERMISO DE HACER CAMBIOS
+    if (usuario.id !== usuarioEncontrado._id.toString()) {
+      return res
+        .status(403)
+        .json({ msg: "No tienes permiso para editar este perfil" });
+    }
+
+    //* ELIMINAR EL USUARIO DE LA BD
     await usuarioEncontrado.deleteOne();
 
     return res.status(200).json({ msg: "Usuario eliminado correctamente" });
